Add a test for the default random number generator.  Updated the logic in there to use a double instead of an int to avoid overflows and negative numbers being returned

+++ src/main/kotlin/com/flexpoker/table/command/aggregate/RandomNumberGenerator.kt
@@ -15,6 +15,6 @@ class DefaultRandomNumberGenerator : RandomNumberGenerator {
     }
 
     override fun pseudoRandomIntBasedOnUUID(uuid: UUID, exclusiveMax: Int): Int {
-        return uuid.leastSignificantBits.absoluteValue.toInt() % exclusiveMax
+        return (uuid.leastSignificantBits.absoluteValue.toDouble() % exclusiveMax).toInt()
     }
 }
